% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_000.R
\name{build_completion_for_package}
\alias{build_completion_for_package}
\title{Provide completion files for the specified R package.}
\usage{
build_completion_for_package(pkgname)
}
\arguments{
\item{pkgname}{The package name.}
}
\description{
Some R packages will not be provided with completion files by
\code{nvimcom} (R.nvim), such as R packages loaded using \code{devtools::load_all}.
This function provides additional completion.
However, it should be noted that the R package does not have an
Rd document, so it is currently not possible to obtain parameter
description for the completion.
}
