% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_000.R
\name{nvim.buildArgsWithoutDescription}
\alias{nvim.buildArgsWithoutDescription}
\title{Build completion of augments without description}
\usage{
nvim.buildArgsWithoutDescription(afile, pkg)
}
\arguments{
\item{afile}{The file to output.}

\item{pkg}{The package name.}
}
\description{
The R package does not have an Rd file. Therefore, the
completion does not involve the explanation of parameters.
}
